services/flagid/requirements.txt
pymongo
requests
python-dotenv
